SM-1493 Move camera-related custom views to the infra module
