+++ feature/external-credential/src/main/res/layout-land/fragment_external_credential_scan_ocr.xml
@@ -12,7 +12,7 @@
         android:scaleType="fitCenter"
         tools:background="@tools:sample/backgrounds/scenic" />
 
-    <com.simprints.infra.view.imagecapture.CaptureMaskView
+    <com.simprints.infra.camera.view.CaptureMaskView
         android:id="@+id/viewfinderMask"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -38,7 +38,7 @@
         tools:text="Place document code in window"
         tools:visibility="visible" />
 
-    <com.simprints.infra.view.imagecapture.CaptureTargetView
+    <com.simprints.infra.camera.view.CaptureTargetView
         android:id="@+id/documentScannerArea"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -55,7 +55,7 @@
         app:strokeWidth="4dp"
         tools:visibility="visible" />
 
-    <com.simprints.infra.view.imagecapture.CaptureProgressView
+    <com.simprints.infra.camera.view.CaptureProgressView
         android:id="@+id/captureProgress"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -119,7 +119,7 @@
 
     </LinearLayout>
 
-    <com.simprints.infra.view.PermissionRequestView
+    <com.simprints.infra.camera.view.PermissionRequestView
         android:id="@+id/permissionRequestView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
