+++ feature/external-credential/src/main/res/layout/fragment_external_credential_scan_ocr.xml
@@ -12,7 +12,7 @@
         android:scaleType="fitCenter"
         tools:background="@tools:sample/backgrounds/scenic" />
 
-    <com.simprints.infra.view.imagecapture.CaptureMaskView
+    <com.simprints.infra.camera.view.CaptureMaskView
         android:id="@+id/viewfinderMask"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -23,7 +23,7 @@
         app:targetViewId="@+id/documentScannerArea"
         tools:visibility="visible" />
 
-    <com.simprints.infra.view.imagecapture.CaptureTargetView
+    <com.simprints.infra.camera.view.CaptureTargetView
         android:id="@+id/documentScannerArea"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -40,7 +40,7 @@
         app:strokeWidth="4dp"
         tools:visibility="visible" />
 
-    <com.simprints.infra.view.imagecapture.CaptureProgressView
+    <com.simprints.infra.camera.view.CaptureProgressView
         android:id="@+id/captureProgress"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -70,11 +70,11 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:gravity="center"
+        android:paddingHorizontal="@dimen/padding_default"
+        android:paddingVertical="4dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/documentScannerArea"
-        android:paddingHorizontal="@dimen/padding_default"
-        android:paddingVertical="4dp"
         tools:text="Image is too bright"
         tools:textColor="@color/simprints_text_white" />
 
@@ -83,8 +83,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:paddingBottom="@dimen/padding_default"
         android:paddingHorizontal="@dimen/padding_default"
+        android:paddingBottom="@dimen/padding_default"
         android:visibility="invisible"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -99,11 +99,11 @@
                 android:id="@+id/iconScanComplete"
                 android:layout_width="36dp"
                 android:layout_height="36dp"
-                android:alpha="0"
-                tools:alpha="1"
                 android:layout_gravity="center"
+                android:alpha="0"
                 android:importantForAccessibility="no"
-                android:src="@drawable/ic_checked_green_large"/>
+                android:src="@drawable/ic_checked_green_large"
+                tools:alpha="1" />
         </FrameLayout>
 
         <TextView
@@ -117,7 +117,7 @@
 
     </LinearLayout>
 
-    <com.simprints.infra.view.PermissionRequestView
+    <com.simprints.infra.camera.view.PermissionRequestView
         android:id="@+id/permissionRequestView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
