+++ feature/external-credential/src/main/res/layout/fragment_external_credential_scan_qr.xml
@@ -11,7 +11,7 @@
         android:layout_height="match_parent"
         tools:visibility="gone" />
 
-    <com.simprints.infra.view.imagecapture.CaptureMaskView
+    <com.simprints.infra.camera.view.CaptureMaskView
         android:id="@+id/viewfinderMask"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -33,7 +33,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"/>
 
-    <com.simprints.infra.view.imagecapture.CaptureTargetView
+    <com.simprints.infra.camera.view.CaptureTargetView
         android:id="@+id/qrScannerArea"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -86,7 +86,7 @@
         app:layout_constraintStart_toStartOf="parent"
         tools:text="No QR code detected" />
 
-    <com.simprints.infra.view.PermissionRequestView
+    <com.simprints.infra.camera.view.PermissionRequestView
         android:id="@+id/permissionRequestView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
