+++ infra/camera/src/main/java/com/simprints/infra/camera/view/CaptureMaskView.kt
@@ -1,4 +1,4 @@
-package com.simprints.infra.view.imagecapture
+package com.simprints.infra.camera.view
 
 import android.content.Context
 import android.graphics.Canvas
@@ -11,8 +11,8 @@ import android.util.AttributeSet
 import android.view.View
 import androidx.annotation.ColorInt
 import androidx.core.content.ContextCompat
-import com.simprints.infra.uibase.R
-import com.simprints.infra.uibase.annotations.ExcludedFromGeneratedTestCoverageReports
+import com.simprints.core.ExcludedFromGeneratedTestCoverageReports
+import com.simprints.infra.camera.R
 import com.simprints.infra.resources.R as IDR
 
 @ExcludedFromGeneratedTestCoverageReports("UI Code")
