+++ infra/camera/src/main/java/com/simprints/infra/camera/view/PermissionRequestView.kt
@@ -1,4 +1,4 @@
-package com.simprints.infra.view
+package com.simprints.infra.camera.view
 
 import android.content.Context
 import android.util.AttributeSet
@@ -7,16 +7,15 @@ import android.widget.Button
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.annotation.StringRes
-import com.simprints.infra.uibase.R
-import com.simprints.infra.uibase.annotations.ExcludedFromGeneratedTestCoverageReports
+import com.simprints.core.ExcludedFromGeneratedTestCoverageReports
+import com.simprints.infra.camera.R
 
 @ExcludedFromGeneratedTestCoverageReports("UI Code")
 class PermissionRequestView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
-    defStyle: Int = 0
+    defStyle: Int = 0,
 ) : LinearLayout(context, attrs, defStyle) {
-
     private val title: TextView
     private val body: TextView
     private val actionButton: Button
@@ -30,21 +29,36 @@ class PermissionRequestView @JvmOverloads constructor(
         actionButton = findViewById(R.id.permissionAction)
     }
 
-    fun init(title: String, body: String, buttonText: String, onClickListener: OnClickListener) {
+    fun init(
+        title: String,
+        body: String,
+        buttonText: String,
+        onClickListener: OnClickListener,
+    ) {
         this.title.text = title
         this.body.text = body
         this.actionButton.text = buttonText
         this.actionButton.setOnClickListener(onClickListener)
     }
 
-    fun init(@StringRes title: Int, @StringRes body: Int, @StringRes buttonText: Int, onClickListener: OnClickListener) {
+    fun init(
+        @StringRes title: Int,
+        @StringRes body: Int,
+        @StringRes buttonText: Int,
+        onClickListener: OnClickListener,
+    ) {
         this.title.setText(title)
         this.body.setText(body)
         this.actionButton.setText(buttonText)
         this.actionButton.setOnClickListener(onClickListener)
     }
 
-    fun init(@StringRes title: Int, body: String, @StringRes buttonText: Int, onClickListener: OnClickListener) {
+    fun init(
+        @StringRes title: Int,
+        body: String,
+        @StringRes buttonText: Int,
+        onClickListener: OnClickListener,
+    ) {
         this.title.setText(title)
         this.body.setText(body)
         this.actionButton.setText(buttonText)
