+++ infra/camera/src/main/res/layout/view_camera_permission_request.xml
