+++ infra/camera/src/main/res/values/attrs.xml
